Change DBClient setup to make multiple clients possible
