+++ app/src/adapters/db/flask_db.py
@@ -8,7 +8,7 @@ Example:
     import src.adapters.db as db
     import src.adapters.db.flask_db as flask_db
 
-    db_client = db.init()
+    db_client = db.init(PostgresEngine())
     app = APIFlask(__name__)
     flask_db.init_app(db_client, app)
 
@@ -46,24 +46,31 @@ from flask import Flask, current_app
 import src.adapters.db as db
 from src.adapters.db.client import DBClient
 
-_FLASK_EXTENSION_KEY = "db"
+_DEFAULT_FLASK_EXTENSION_KEY = "db"
 
 
-def init_app(db_client: DBClient, app: Flask) -> None:
+def init_app(
+    db_client: DBClient, app: Flask, client_name: str = _DEFAULT_FLASK_EXTENSION_KEY
+) -> None:
     """Initialize the Flask app.
 
     Add the database to the Flask app's extensions so that it can be
     accessed by request handlers using the current app context.
 
+    If you use multiple DB clients, you can differentiate them by
+    specifying a client_name.
+
     see get_db
     """
-    app.extensions[_FLASK_EXTENSION_KEY] = db_client
+    app.extensions[client_name] = db_client
 
 
-def get_db(app: Flask) -> DBClient:
+def get_db(app: Flask, client_name: str = _DEFAULT_FLASK_EXTENSION_KEY) -> DBClient:
     """Get the database connection for the given Flask app.
 
     Use this in request handlers to access the database from the active Flask app.
+
+    Specify the same client_name as used in init_app to get the correct client
 
     Example:
         from flask import current_app
@@ -73,14 +80,16 @@ def get_db(app: Flask) -> DBClient:
         def health():
             db_client = flask_db.get_db(current_app)
     """
-    return app.extensions[_FLASK_EXTENSION_KEY]
+    return app.extensions[client_name]
 
 
 P = ParamSpec("P")
 T = TypeVar("T")
 
 
-def with_db_session(f: Callable[Concatenate[db.Session, P], T]) -> Callable[P, T]:
+def with_db_session(
+    *, client_name: str = _DEFAULT_FLASK_EXTENSION_KEY
+) -> Callable[[Callable[Concatenate[db.Session, P], T]], Callable[P, T]]:
     """Decorator for functions that need a database session.
 
     This decorator will create a new session object and pass it to the function
@@ -88,18 +97,25 @@ def with_db_session(f: Callable[Concatenate[db.Session, P], T]) -> Callable[P, T
     To start a transaction use db_session.begin()
 
     Usage:
-        @with_db_session
+        @with_db_session()
         def foo(db_session: db.Session):
             ...
 
-        @with_db_session
+        @with_db_session()
         def bar(db_session: db.Session, x, y):
+            ...
+
+        @with_db_session(client_name="not_the_primary_client")
+        def fiz(db_session: db.Session, x, y, z):
             ...
     """
 
-    @wraps(f)
-    def wrapper(*args: Any, **kwargs: Any) -> T:
-        with get_db(current_app).get_session() as session:
-            return f(session, *args, **kwargs)
+    def decorator_func(f: Callable[Concatenate[db.Session, P], T]) -> Callable[P, T]:
+        @wraps(f)
+        def wrapper(*args: Any, **kwargs: Any) -> T:
+            with get_db(current_app, client_name=client_name).get_session() as session:
+                return f(session, *args, **kwargs)
 
-    return wrapper
+        return wrapper
+
+    return decorator_func
