+++ app/tests/src/adapters/db/test_flask_db.py
@@ -4,6 +4,7 @@ from sqlalchemy import text
 
 import src.adapters.db as db
 import src.adapters.db.flask_db as flask_db
+from src.adapters.db.engine.postgres_engine import PostgresEngine
 
 
 # Define an isolated example Flask app fixture specific to this test module
@@ -11,7 +12,7 @@ import src.adapters.db.flask_db as flask_db
 @pytest.fixture
 def example_app() -> Flask:
     app = Flask(__name__)
-    db_client = db.init()
+    db_client = db.init(PostgresEngine())
     flask_db.init_app(db_client, app)
     return app
 
@@ -28,7 +29,21 @@ def test_get_db(example_app: Flask):
 
 def test_with_db_session(example_app: Flask):
     @example_app.route("/hello")
-    @flask_db.with_db_session
+    @flask_db.with_db_session()
+    def hello(db_session: db.Session):
+        with db_session.begin():
+            return {"data": db_session.scalar(text("SELECT 'hello, world'"))}
+
+    response = example_app.test_client().get("/hello")
+    assert response.get_json() == {"data": "hello, world"}
+
+
+def test_with_db_session_not_default_name(example_app: Flask):
+    db_client = db.init(PostgresEngine())
+    flask_db.init_app(db_client, example_app, client_name="something_else")
+
+    @example_app.route("/hello")
+    @flask_db.with_db_session(client_name="something_else")
     def hello(db_session: db.Session):
         with db_session.begin():
             return {"data": db_session.scalar(text("SELECT 'hello, world'"))}
